fix the tests and move audit utils to back office
