+++ packages/utils/package.json
@@ -2,6 +2,7 @@
   "name": "@mcw/utils",
   "version": "0.0.0",
   "private": true,
+  "type": "module",
   "sideEffects": false,
   "exports": {
     ".": "./src/index.ts"
